refactoring

+++ c_server/inc/server.h
@@ -1,3 +1,3 @@
 int init_server();
-void connection_handler(float * H, float * T, int lamp[], int ac[], int sp[], int so[]);
+void connection_handler(double * H, double * T, int lamp[], int ac[], int sp[], int so[]);
 void close_sockets();
